// src/entities/User/model/types/user.ts
import { UserRole } from '../consts/consts';
import { JsonSettings } from './jsonSettings';

export interface User {
    id: string;
    username: string;
    avatar?: string;
    roles?: Array<UserRole>;
    jsonSettings?: JsonSettings;
}

export interface UserSchema {
    authData?: User;
    _mounted?: boolean;
}